chore: rename from passkey to auth
