+++ prisma/schema.prisma
@@ -9,23 +9,25 @@ datasource db {
 }
 
 model Merchant {
-  id                  Int             @id @default(autoincrement())
-  name                String          @db.VarChar(255)
-  surname             String          @db.VarChar(255)
-  email               String          @unique
+  id                  Int                @id @default(autoincrement())
+  name                String             @db.VarChar(255)
+  surname             String             @db.VarChar(255)
+  email               String             @unique
+  rainId              String?
   tokenAuth           String?
-  phoneNumber         String?         @unique
+  phoneNumber         String?            @unique
   companyNumber       String?
-  companyJurisdiction String?         @db.Char(2)
-  fee                 Decimal         @default(6.50) @db.Decimal(4, 2)
-  walletAddress       String          @unique
-  createdAt           DateTime        @default(now())
-  updatedAt           DateTime        @updatedAt
+  companyJurisdiction String?            @db.Char(2)
+  fee                 Decimal            @default(6.50) @db.Decimal(4, 2)
+  walletAddress       String             @unique
+  createdAt           DateTime           @default(now())
+  updatedAt           DateTime           @updatedAt
   addressId           BigInt
   Buyer               Buyer[]
   Compliance          Compliance[]
-  registeredAddress   PhysicalAddress @relation(fields: [addressId], references: [id])
+  registeredAddress   PhysicalAddress    @relation(fields: [addressId], references: [id])
   Order               Transaction[]
+  employees           MerchantEmployee[]
 
   @@unique([companyNumber, companyJurisdiction])
 }
@@ -94,28 +96,9 @@ model PhysicalAddress {
   Transaction Transaction[]
 }
 
-model RainCompany {
-  id          Int           @id @default(autoincrement())
-  ref         String        @unique
-  name        String
-  createdAt   DateTime      @default(now())
-  updatedAt   DateTime      @updatedAt
-  rainMembers RainMembers[]
-}
-
-model RainMembers {
-  id          Int         @id @default(autoincrement())
-  active      Boolean     @default(true)
-  userId      Int
-  companyId   Int
-  createdAt   DateTime    @default(now())
-  rainCompany RainCompany @relation(fields: [companyId], references: [id])
-  user        User        @relation(fields: [userId], references: [id])
-}
-
 model RegisteredPasskey {
   id           Int                 @id @default(autoincrement())
-  name         String?              @default("") @db.VarChar(255)
+  name         String?             @default("") @db.VarChar(255)
   credentialId String              @unique
   publicKey    String
   algorithm    CredentialAlgorithm @default(ES256)
@@ -123,27 +106,28 @@ model RegisteredPasskey {
   createdAt    DateTime            @default(now())
   updatedAt    DateTime            @updatedAt
   user         User                @relation(fields: [userId], references: [id])
-  session      Session[]
 }
 
-model Session {
-  id                Int               @id @default(autoincrement())
-  token             String            @unique
-  deviceId          Int
-  createdAt         DateTime          @default(now())
-  updatedAt         DateTime          @updatedAt
-  registeredPasskey RegisteredPasskey @relation(fields: [deviceId], references: [id])
+model MerchantEmployee {
+  id         Int      @id @default(autoincrement())
+  rain       String?  @unique @db.VarChar(36)
+  owner      Boolean  @default(false)
+  merchantId Int
+  userId     Int
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
+  user       User     @relation(fields: [userId], references: [id])
+  merchant   Merchant @relation(fields: [merchantId], references: [id])
 }
 
 model User {
   id                Int                 @id @default(autoincrement())
   email             String              @unique
   username          String              @unique @db.VarChar(15)
-  rain              String?             @unique @db.VarChar(36)
   createdAt         DateTime            @default(now())
   updatedAt         DateTime            @updatedAt
-  rainMembers       RainMembers[]
   registeredPasskey RegisteredPasskey[]
+  merchantEmployee  MerchantEmployee[]
 }
 
 enum TransactionStatus {
