+++ src/index.ts
@@ -8,7 +8,7 @@ import rateLimit from '@fastify/rate-limit';
 import jwt from '@fastify/jwt';
 import rawBody from 'fastify-raw-body';
 
-import { Home, Merchant, Bridge, Passkey } from './v1/routes/index';
+import { Home, Merchant, Bridge, Auth } from './v1/routes/index';
 import { Config } from './config';
 
 const startServer = async () => {
@@ -43,7 +43,7 @@ const startServer = async () => {
       .register(Home)
       .register(Merchant, { prefix: '/v1/merchant' })
       .register(Bridge, { prefix: '/v1/bridge' })
-      .register(Passkey, { prefix: '/v1/passkey' });
+      .register(Auth, { prefix: '/v1/auth' });
 
     const serverOptions = {
       port: Config.port,
